Yellow/requirements.txt
args==0.1.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
clint==0.5.1
idna==2.10
numpy==1.19.1
pyfiglet==0.8.post1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tinydb==4.1.1
urllib3==1.25.10
